requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pandas
numpy
matplotlib
seaborn
plotly
aiohttp
python-dotenv
pyjwt
requests
psutil
ipython
jupyter
kaleido

# Computer Vision Libraries
mediapipe
cvzone
opencv-python
opencv-contrib-python

# Streamlit for interactive apps
streamlit
streamlit-webrtc

# Additional image/video processing
pillow
imageio
scikit-image

# ZeroMQ for distributed kernel workers
pyzmq
redis
